refactor: Standardize import, imposing mild conventions
https://github.com/qiboteam/qibolab/issues/949#issue-2448262112
