+++ src/qibolab/_core/instruments/dummy.py
@@ -2,19 +2,18 @@ import numpy as np
 from pydantic import Field
 from qibo.config import log
 
-from qibolab._core.components.channels import Channel
+from qibolab._core.components import Channel, Config
 from qibolab._core.execution_parameters import (
     AcquisitionType,
     AveragingMode,
     ExecutionParameters,
 )
 from qibolab._core.identifier import ChannelId
-from qibolab._core.pulses.pulse import Acquisition
+from qibolab._core.pulses import Acquisition
 from qibolab._core.sequence import PulseSequence
 from qibolab._core.sweeper import ParallelSweepers
 from qibolab._core.unrolling import Bounds
 
-from ..components import Config
 from .abstract import Controller
 from .oscillator import LocalOscillator
